refactor(phase-4-pr4.5): collapse handlers/tests boilerplate per code-review
Self-review feedback + user grill on PR size — cut convention tax.

- Extract `auditActorFrom` + `emitAudit` to `audit-helpers.ts`. Six
  handlers stop repeating the 3-line actor triple and the 11-field audit
  struct.
- Consolidate 5 namespace mutation handlers into a single
  `namespace-mutations.ts` module (was 5 separate files with duplicated
  imports + docstrings). Test files mirror with one
  `namespace-mutations.test.ts` carrying 5 describe blocks.
- Promote the in-memory `NamespaceRepository` test double to
  `@mediforce/platform-api/testing`. Removes 5 inline copies across
  handler tests + the api-integration loopback (each ~80 LOC stub).

Net: -1385 / +38 LOC (≈30% PR reduction). Tests still pass (2310/2311;
api-plugins-route flake is pre-existing parallel timeout — passes
isolated).
